Add cmake as additional build system
cmake is a build system that is widely used
and should be added as an alternative for meson.
It is cross platform but currently there are hardcoded
paths which only work for Linux.
Also it uses pkg-config for finding installed packages.
And the po translations are not yet implemented.
So there is definitely work to do until it offers the
same functionality as the meson build scripts.