Extend keyboard support for search and fullscreen
This patch allows to search in the channels, favourites, or games views,
just by starting to type. The search bar can be toggled by pressing
Ctrl+F or dismissed by pressing Escape.

Pressing Escape in the player view exits fullscreen or closes the player
and returns to the browse views if windowed. The F key can be used to
toggle fullscreen.

Ctrl+Q immediately quits the application.