Update game box art cache in the background.
In case of a cache hit, the picture is loaded from disk and an update
task is scheduled. The task uses an HTTP HEAD request to check if the
cache entry must be updated. If it does, the relevant GtGame object is
notified.

Note that there is no maximum lifespan for cache entries. Also, the
downloaded pictures are not processed before being saved to disk.

Credit to Tim Dengel for the suggestion.