Replace unsafe format string by safe alternative
Compiling with -Werror=format-security fails because a printf-like function was used in an unsafe way.
Format strings for functions like this should be string literals, to avoid injection of
format specifiers, which can lead to undefined behavior if they are not filled in with arguments.

If one wants to pass a normal string to one of those functions without formatting, the format string
"%s" should be used, with the desired string passed as an argument for it.

See also:
https://stackoverflow.com/questions/17260409/fprintf-error-format-not-a-string-literal-and-no-format-arguments-werror-for
https://gcc.gnu.org/onlinedocs/gcc/Warning-Options.html